moved content of FastTrackDeDxProducer.h to .cc and deleted the former
